everything worked

+++ app/src/main/java/com/example/goodweather/domain/entity/Sys.kt
@@ -2,9 +2,9 @@ package com.example.goodweather.domain.entity
 
 data class Sys(
     val type : Int,
-    val id : Int,
-    val message : Int,
+    val id : Long,
+    val message : Double,
     val country : String,
-    val sunrise : Int,
-    val sunset : Int
+    val sunrise : Long,
+    val sunset : Long
 )
