+++ app/src/main/java/com/example/goodweather/domain/entity/Wind.kt
@@ -1,6 +1,6 @@
 package com.example.goodweather.domain.entity
 
 data class Wind(
-    val speed : Int,
-    val deg : Int
+    val speed : Double,
+    val deg : Double
 )
